Add rubyonremote.com
https://rubyonremote.com/

Hi,
This list has been of huge help for me personally while getting into remote work. When I was looking for ruby jobs i had to go through a lot of job boards to find them so I decided to create an easy to use clean UI for remote ruby jobs. I also wanted to make sure we have a high quality of remote jobs and not quantity because ensuring ruby as a required skill was a must for me. So I curate the job board from the web manually every day to avoid having jobs which don't use ruby.

It would be great if you could add my site to the list! Cheers.